pyproject.toml
[project]
name = "django_app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adrf>=0.1.12",
    "aiohttp>=3.13.5",
    "daphne>=4.2.1",
    "django>=6.0.4",
    "django-crispy-forms>=2.6",
    "django-ninja>=1.6.2",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "mysqlclient>=2.2.8",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "structlog>=25.5.0",
    "uvicorn>=0.45.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "server.settings"
testpaths = ["tests"]
